% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codes.R
\name{aggregate_pers}
\alias{aggregate_pers}
\title{Aggregate category percentages in groups}
\usage{
aggregate_pers(data, groups = v5_v4_aggregation_relations(), na.rm = FALSE,
  keep = FALSE)
}
\arguments{
\item{data}{dataset to use in aggregation}

\item{groups}{(named) list of variable name vectors to aggregate to a new one
(as given in the name); see default value for an example of the format}

\item{na.rm}{passed on to \code{\link{sum}}}

\item{keep}{keep variables that were aggregated in result?}
}
\description{
General function to aggregate percentage variables by creating a new
variable holding the sum. If a variable with the name for the aggregate
already exists, it is overwritten, giving a warning if it is changed, not NA,
not zero and not named "peruncod".
}

