% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{aggregate_pers_cee}
\alias{aggregate_pers_cee}
\title{Aggregate cee-categories to main categories}
\usage{
aggregate_pers_cee(data)
}
\arguments{
\item{data}{dataset to use in aggregation}
}
\description{
Adds the code frequencies in a dataset of the 4 digit per-variables (per1011 to per7062 - mostly used in codings
of Central and Eastern European countries) to the main categories in the coding scheme (3 digits).
}
\details{
A wrapper of \code{\link{aggregate_pers}} using \code{cee_aggregation_relations}.
}
\seealso{
\code{\link{aggregate_pers}}
}
