% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{attach_year}
\alias{attach_year}
\title{Compute year from date variable in MPDS}
\usage{
attach_year(mpds)
}
\arguments{
\item{mpds}{a dataframe in format of Manifesto Project Main Dataset}
}
\value{
input data with year variable attached
}
\description{
Compute year from date variable in MPDS
}

