% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{clarity_dimensions}
\alias{clarity_dimensions}
\title{Default programmatic clarity dimensions from 
Giebler/Lacewell/Regel/Werner 2015.}
\usage{
clarity_dimensions()
}
\description{
Default programmatic clarity dimensions from 
Giebler/Lacewell/Regel/Werner 2015.
}
\references{
Giebler/Lacewell/Regel/Werner (2015). Mass, Catch-all, or 
Programmatic? Toward an Empirical Classification of Party Types. Manuscript.
}
