% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/codes.R
\name{aggregate_cee_codes}
\alias{aggregate_cee_codes}
\alias{aggregate_cee_codes.ManifestoCorpus}
\alias{aggregate_cee_codes.ManifestoDocument}
\alias{aggregate_cee_codes.default}
\title{Process CMP codings}
\usage{
aggregate_cee_codes(x)

\method{aggregate_cee_codes}{default}(x)

\method{aggregate_cee_codes}{ManifestoDocument}(x)

\method{aggregate_cee_codes}{ManifestoCorpus}(x)
}
\arguments{
\item{x}{Vector of codes, ManifestoDocument or ManifestoCorpus}
}
\description{
Several functions to process the CMP codings
}
\details{
\code{aggregate_cee_codes} Aggregates the sub-categories used
in coding several manifestos in Central and Eastern Europe (4 digits) to
the main categories in the coding scheme (3 digits).
}

