% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db_api.R
\name{formatmpds}
\alias{formatmpds}
\title{Format the main data set}
\usage{
formatmpds(mpds)
}
\arguments{
\item{mpds}{A data.frame with a main data set version to be formatted}
}
\description{
Creates the format that is visible to the R user
from the internal data.frames files (in cache or from the API)
}

