% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_api.R
\name{formatpartiesds}
\alias{formatpartiesds}
\title{Format the parties data set}
\usage{
formatpartiesds(partiesds)
}
\arguments{
\item{partiesds}{A data.frame with a parties dataset to be formatted}
}
\description{
Creates the format that is visible to the R user
from the internal data.frames files (in cache or from the API)
}
