% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_api.R
\name{get_mpdb}
\alias{get_mpdb}
\title{Download content from the Manifesto Database}
\usage{
get_mpdb(type, parameters = c(), versionid = NULL, apikey = NULL)
}
\arguments{
\item{type}{string of \code{"meta", "text", "original", "main", "versions"} 
to indicate type of content to get}

\item{parameters}{content filter parameters specific to type}

\item{versionid}{character string specifying the corpus version to use, either
a name or tag as in the respective columns of the value of
\code{\link{mp_corpusversions}} and the API}

\item{apikey}{API key to use, defaults to \code{NULL}, which means the key 
currently stored in the variable \code{apikey} of the
environment \code{mp_globalenv} is used.}
}
\description{
Internal implementation. For more convenient access and caching use one of 
\code{\link{mp_corpus}}, 
\code{\link{mp_availability}},  
\code{\link{mp_maindataset}}.
}
