% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{get_viacache}
\alias{get_viacache}
\title{Get API results via cache}
\usage{
get_viacache(type, ids = c(), cache = TRUE, versionid = NULL, ...)
}
\arguments{
\item{type}{type of objects to get (metadata, documents, ...) as a string. Types
are defined as constants in globals.R}

\item{ids}{identifiers of objects to get. Depending on the type a data.frame or vector of identifiers.}

\item{cache}{whether to use (TRUE) or bypass (FALSE) cache, defaults to TRUE}

\item{versionid}{string identifier of version to use}

\item{...}{additional parameters handed over to get_mpdb}
}
\description{
Get API results via cache
}
\details{
This function is internal to manifestoR and not designed for use from
other namespaces
}

