% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\docType{class}
\name{ManifestoCorpus}
\alias{ManifestoCorpus}
\title{Manifesto Corpus class}
\usage{
ManifestoCorpus(csource = ManifestoJSONSource())
}
\arguments{
\item{csource}{a \code{\link{ManifestoJSONSource}}, see \code{\link[tm]{Source}}}
}
\description{
Objects of this class are returned by \code{\link{mp_corpus}}.
}
\details{
A \code{tm} \code{\link[tm]{Corpus}} storing \code{\link{ManifestoDocument}s}


For usage and structure of the stored documents see
\code{\link{ManifestoDocument}}.
}
\examples{
\dontrun{corpus <- mp_corpus(subset(mp_maindataset(), countryname == "Russia"))}
}
