% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\docType{class}
\name{ManifestoDocumentMeta}
\alias{ManifestoDocumentMeta}
\title{Manifesto Document Metadata}
\usage{
ManifestoDocumentMeta(meta = list(), id = character(0))
}
\arguments{
\item{meta}{a named list with tag-value pairs of document meta information}

\item{id}{a character giving a unique identifier for the text document}
}
\description{
Manifesto Document Metadata
}

