% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestoR-package.r
\docType{package}
\name{manifestoR-package}
\alias{manifestoR}
\alias{manifestoR-package}
\title{manifestoR: Access and Process Data and Documents of the Manifesto Project}
\description{
Provides access to coded election programmes from the Manifesto Corpus and to the Manifesto Project's Main Dataset and routines to analyse this data. The Manifesto Project \url{https://manifesto-project.wzb.eu} collects and analyses election programmes across time and space to measure the political preferences of parties. The Manifesto Corpus contains the collected and annotated election programmes in the Corpus format of the package 'tm' to enable easy use of text processing and text mining functionality. Specific functions for scaling of coded political texts are included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ManifestoProject/manifestoR}
  \item \url{https://manifesto-project.wzb.eu/}
  \item Report bugs at \url{https://github.com/ManifestoProject/manifestoR/issues}
}

}
\author{
\strong{Maintainer}: Sven Regel \email{sven.regel@wzb.eu}

Authors:
\itemize{
  \item Jirka Lewandowski \email{jirka.lewandowski@wzb.eu}
  \item Nicolas Merz \email{nicolas.merz@wzb.eu}
}

Other contributors:
\itemize{
  \item Pola Lehmann \email{pola.lehmann@wzb.eu} [contributor]
  \item Paul Muscat \email{paul.muscat@wzb.eu} [contributor]
}

}
\keyword{internal}
