% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.R
\docType{class}
\name{ManifestoSource}
\alias{ManifestoSource}
\title{Data Source for Manifesto Corpus}
\usage{
ManifestoSource(texts)
}
\arguments{
\item{texts}{texts of the manifesto documents}
}
\description{
Data Source for Manifesto Corpus
}
\details{
Used internally for constructing \code{\link{ManifestoCorpus}} objects.
}

