% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto.R
\name{mp_cite}
\alias{mp_cite}
\title{Print Manifesto Corpus citation information}
\usage{
mp_cite(corpus_version = mp_which_corpus_version(),
  core_versions = mp_which_dataset_versions(), apikey = NULL)
}
\arguments{
\item{corpus_version}{corpus version for which citation should be printed}

\item{core_versions}{core version for which citation should be printed}

\item{apikey}{API key to use. Defaults to \code{NULL}, resulting in using
the API key set via \code{\link{mp_setapikey}}.}
}
\description{
Print Manifesto Corpus citation information
}
