% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto.R
\name{mp_coreversions}
\alias{mp_coreversions}
\title{List the available versions of the Manifesto Project's Main Dataset}
\usage{
mp_coreversions(apikey = NULL, cache = TRUE)
}
\arguments{
\item{apikey}{API key to use. Defaults to \code{NULL}, resulting in using
the API key set via \code{\link{mp_setapikey}}.}

\item{cache}{Boolean flag indicating whether to use locally cached data if
available.}
}
\description{
List the available versions of the Manifesto Project's Main Dataset
}
\details{
For the available versions of the corpus, see \code{\link{mp_corpusversions}}
}
\examples{
\dontrun{mp_coreversions()}
}

