% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedication.R
\name{mp_dedication}
\alias{mp_dedication}
\title{Print manifestoR package dedication}
\usage{
mp_dedication()
}
\value{
\code{mp_dedication} returns the package dedication
}
\description{
Print manifestoR package dedication
}
