% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{mp_load_cache}
\alias{mp_load_cache}
\title{Load manifestoR's cache}
\usage{
mp_load_cache(cache = NULL, file = "mp_cache.RData")
}
\arguments{
\item{cache}{an environment that should function as manifestoR's new cache.
If this is NULL, the environment is loaded from the file specified by argument file.}

\item{file}{a file name from where the cache environment should be loaded}
}
\description{
Load a cache from a variable or file to manifestoR's current working
environment.
}
\examples{
\dontrun{mp_load_cache() ## loads cache from file "mp_cache.RData"}
}
