% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{mp_save_cache}
\alias{mp_save_cache}
\title{Save manifestoR's cache}
\usage{
mp_save_cache(file = "mp_cache.RData")
}
\arguments{
\item{file}{a file from which to load the cache environment}
}
\description{
Saves manifestoR's cache to the file system. This function can and should be
used to store downloaded snapshots of the Manifesto Project Corpus Database
to your local hard drive. They can then be loaded via \code{\link{mp_load_cache}}.
Caching data in the file system ensures reproducibility of the scripts and
analyses, enables offline use of the data and reduces unnecessary traffic
and waiting times.
}
\examples{
\dontrun{mp_save_cache() ## save to "mp_cache.RData" in current working directory}
}
