% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_api.R
\name{mp_setapikey}
\alias{mp_setapikey}
\title{Set the API key for the Manifesto Documents Database.}
\usage{
mp_setapikey(key.file = NULL, key = NA)
}
\arguments{
\item{key.file}{file name containing the API key}

\item{key}{new API key}
}
\description{
If you do not have an API key for the Manifesto Documents Database,
you can create one via your profile page on 
\url{https://manifesto-project.wzb.eu}.
If you do not have an account, you can register on the webpage.
}
\details{
The key is read from the file specified in \code{key.file}. If this
argument is \code{NULL}, the key given in the argument \code{key} is used.
}
