% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cache.R
\name{mp_use_corpus_version}
\alias{mp_use_corpus_version}
\title{Use a specific version of the Manifesto Project Corpus}
\usage{
mp_use_corpus_version(versionid, apikey = NULL)
}
\arguments{
\item{versionid}{character id of the version to use (as received from API and
\code{\link{mp_corpusversions}})}

\item{apikey}{API key to use. Defaults to \code{NULL}, resulting in using
the API key set via \code{\link{mp_setapikey}}.}
}
\description{
The internal cache of manifestoR will be updated to the specified version
and all future calls to the API will request for the specified version. Note
that this versioning applies to the corpus' texts and metadata, and not the
versions of the core dataset. For this see \code{\link{mp_coreversions}}
}

