% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_api.R
\name{mpdb_api_request}
\alias{mpdb_api_request}
\title{Manifesto Project DB API request}
\usage{
mpdb_api_request(file, body)
}
\arguments{
\item{file}{file to request below apiroot url}

\item{body}{body text of the posted request: should contain the parameters
as specified by the Manifesto Project Database API}
}
\description{
gets the requested url and passes HTTP header error codes on to raise R
errors with the same text
}

