% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_helpers.R
\name{na_replace}
\alias{na_replace}
\title{Replace NAs in vector with fixed value}
\usage{
na_replace(vec, value = 0L)
}
\arguments{
\item{vec}{vector to replace NAs in}

\item{value}{value to inject for NA}
}
\description{
Replace NAs in vector with fixed value
}
