% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_general.R
\name{null_to_na}
\alias{null_to_na}
\title{Convert NULL to NA}
\usage{
null_to_na(x)
}
\arguments{
\item{x}{element}
}
\value{
NA if the element is NULL, the element otherwise
}
\description{
Convert NULL to NA
}

