% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.R
\name{readManifesto}
\alias{readManifesto}
\title{Reader for \code{\link{ManifestoSource}}}
\usage{
readManifesto(elem, language, id)
}
\arguments{
\item{elem}{a named list with the component \code{content}}

\item{language}{is ignored}

\item{id}{a character giving a unique identifier for the created text document}
}
\description{
Reader for \code{\link{ManifestoSource}}
}
\details{
Used internally for constructing \code{\link{ManifestoCorpus}} objects.
For the general mechanism refer to \code{tm}s \code{\link[tm]{Reader}}
documentation.
}

