% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-NLP.R, R/reexport-tm.R,
%   R/reexport-tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{content}
\alias{content<-}
\alias{meta}
\alias{meta<-}
\alias{tm_map}
\alias{getElem}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{NLP}{\code{\link[NLP:generics]{content}}, \code{\link[NLP:generics]{content<-}}, \code{\link[NLP:generics]{meta}}, \code{\link[NLP:generics]{meta<-}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tm}{\code{\link[tm:Source]{getElem}}, \code{\link[tm]{tm_map}}}
}}

