% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/renaming.R
\name{rename_to_v9}
\alias{rename_files_to_v9}
\alias{rename_to_v9}
\title{Rename function names to to manifestoR v0.9}
\usage{
rename_to_v9(infile, outfile = infile)

rename_files_to_v9(pattern = "*.R((md)|(markdown))?$", recursive = TRUE,
  ...)
}
\arguments{
\item{infile}{name of file with old script}

\item{outfile}{name of destination for new script, defaults to overwrite the old one}

\item{pattern}{pattern of file names, defaults to R and R markdown files}

\item{recursive}{passen on to \code{\link{list.files}}}

\item{...}{more arguments passed on to \code{\link{list.files}}}
}
\value{
success flag

success flag
}
\description{
Rename function names to to manifestoR v0.9

Rename function names to to manifestoR v0.9
}

