% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaling_general.R
\name{rep.data.frame}
\alias{rep.data.frame}
\title{Replicates cases in a data.frame}
\usage{
\method{rep}{data.frame}(x, times = 1, ...)
}
\arguments{
\item{x}{data.frame to replicate}

\item{times}{number of replications}

\item{...}{unused}
}
\value{
data.frame with cases replicated
}
\description{
Replicates cases in a data.frame
}

