% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scaling_general.R
\name{rescale}
\alias{rescale}
\title{Simple linear rescaling of positions}
\usage{
rescale(pos, newmin = -1, newmax = 1, oldmin = min(pos),
  oldmax = max(pos))
}
\arguments{
\item{pos}{position data to be rescaled}

\item{newmin}{indicates the minimum of the new scale (default is -1)}

\item{newmax}{indicates the maximum of the new scale (default is +1)}

\item{oldmin}{indicates the minimum of the existing scale. Can be used to rescale from a known theoretical scale (e.g. -100). If left empty the empirical minimum is used.}

\item{oldmax}{indicates the maximum of the existing. See above.}
}
\description{
Simple linear rescaling of positions
}

