% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_rile.R
\name{rile}
\alias{rile}
\alias{logit_rile}
\title{RILE}
\usage{
rile(x)

logit_rile(x)
}
\arguments{
\item{x}{A data.frame with cases to be scaled, variables named "per..."}
}
\description{
Computes the RILE or other bipolar linear scaling measures for each case in a
data.frame or ManifestoCorpus or for a ManifestoDocument
}
