% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{aggregate_pers}
\alias{aggregate_pers}
\title{Aggregate category percentages in groups}
\usage{
aggregate_pers(
  data,
  groups = v5_v4_aggregation_relations(),
  na.rm = FALSE,
  keep = FALSE,
  overwrite = names(groups),
  verbose = TRUE
)
}
\arguments{
\item{data}{dataset to use in aggregation}

\item{groups}{(named) list of variable name vectors to aggregate to a new one
(as given in the name); see default value for an example of the format}

\item{na.rm}{passed on to \code{\link{sum}}}

\item{keep}{keep variables that were aggregated in result?}

\item{overwrite}{Names of the variables that are allowed to be overwritten by
aggregate. Defaults to all aggregate variable names.}

\item{verbose}{show messages in case of possibly problematic side effects}
}
\description{
\code{aggregate_pers} is a general function to aggregate percentage variables by creating a new
variable holding the sum. If a variable with the name for the aggregate
already exists and if it is part of the `overwrite` parameter, it is
overwritten.
}
\seealso{
\code{\link{aggregate_pers_cee}}
}
