% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R, R/scaling_rile.R
\name{v4_categories}
\alias{rile_l}
\alias{rile_r}
\alias{v4_categories}
\alias{v5_v4_aggregation_relations}
\title{Lists of categories and category relations}
\usage{
v4_categories()

v5_v4_aggregation_relations()

rile_r()

rile_l()
}
\description{
Code numbers of the Manifesto Project's category scheme. For documentation
see \url{https://manifesto-project.wzb.eu/datasets}.
}

