% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{aggregate_cee_codes}
\alias{aggregate_cee_codes}
\alias{recode_v5_to_v4}
\title{Process CMP codings}
\usage{
aggregate_cee_codes(x)

recode_v5_to_v4(x)
}
\arguments{
\item{x}{Vector of codes, ManifestoDocument or ManifestoCorpus}
}
\description{
Several functions to process the CMP codings
}
\details{
\code{aggregate_cee_codes} Aggregates the sub-categories used
in coding several manifestos in Central and Eastern Europe (4 digits) to
the main categories in the coding scheme (3 digits).

\code{recode_v5_to_v4} recode the CMP codings according to
the more specialized Coding Handbook Version 5 to the more general
categories of Handbook Version 4. Codes 202.2, 605.2 and 703.2 are
converted to a 000, while all other subcategory codes with an appended
dot and fourth digit are aggregated to the corresponding three-digit 
main category.
}
