% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{codes}
\alias{codes}
\alias{codes.ManifestoDocument}
\alias{codes.ManifestoCorpus}
\alias{codes<-}
\alias{codes<-.ManifestoDocument}
\alias{code_layers}
\title{Access the codes of a Manifesto Document or Corpus}
\usage{
codes(x, layer = "cmp_code")

\method{codes}{ManifestoDocument}(x, layer = "cmp_code")

\method{codes}{ManifestoCorpus}(x, layer = "cmp_code")

codes(x, layer = "cmp_code") <- value

\method{codes}{ManifestoDocument}(x, layer = "cmp_code") <- value

code_layers(x)
}
\arguments{
\item{x}{document or corpus to get the codes from}

\item{layer}{layer of codings to access, defaults to cmp_code, alternative: eu_code}

\item{value}{new codes}
}
\description{
With the accessor the codes of a Manifesto Document can be
read and modified. The codes of a Manifesto Corpus can only be read,
modification needs to be done document-wise.
}
