% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/codes.R
\name{count_codes}
\alias{count_codes}
\title{Count the codings from a ManifestoDocument}
\usage{
count_codes(doc, with_eu_codes = meta(doc, "has_eu_code"), prefix = "per",
  relative = TRUE)
}
\arguments{
\item{doc}{ManifestoDocument to use}

\item{with_eu_codes}{Whether to include special EU code layer; by default taken
from the document's metadata}

\item{prefix}{prefix for naming the count/percentage columns in the resulting data.frame}

\item{relative}{If true, percentages are returned, absolute counts else}
}
\value{
A data.frame with onw row and the counts/percentages as columns
}
\description{
Count the codings from a ManifestoDocument
}

