% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manifesto.R
\docType{class}
\name{ManifestoAvailability}
\alias{ManifestoAvailability}
\title{Manifesto Availability Information class}
\description{
Objects returned by \code{\link{mp_availability}}.
}
\details{
ManifestoAvailability objects are lists with a key \code{query}, containing
the original id set which was queried, and a key \code{availability},
containing information derived from the Manifesto Project's document metadata
database about which types of documents are available for the query.
}
\examples{
\dontrun{
wanted <- data.frame(party=c(41320, 41320), date=c(200909, 200509))
mp_availability(wanted)
}
}

