% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestoR-package.r
\docType{package}
\name{manifestoR}
\alias{manifestoR}
\alias{manifestoR-package}
\title{Access and process data and documents of the Manifesto Project}
\description{
manifestoR R package
}
\details{
Access and process data and documents of the Manifesto Project
\tabular{ll}{
Package: \tab manifestoR\cr
Type: \tab Package\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\references{
\url{https://manifesto-project.wzb.eu/}
}
\author{
Jirka Lewandowski \email{jirka.lewandowski@wzb.eu}
}
