% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cache.R
\name{mp_corpusversions}
\alias{mp_corpusversions}
\title{List the available versions of the Manifesto Project's Corpus}
\usage{
mp_corpusversions(apikey = NULL)
}
\arguments{
\item{apikey}{API key to use. Defaults to \code{NULL}, resulting in using
the API key set via \code{\link{mp_setapikey}}.}
}
\value{
a character vector with the available version ids
}
\description{
The Manifesto Project Database API assigns a new version code whenever changes
to the corpus texts or metadata are made.
}
\details{
This function always bypasses the cache.
}

