% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity.R
\name{mp_rmps}
\alias{mp_rmps}
\title{Relative measure of party size (RMPS)}
\usage{
mp_rmps(data)
}
\arguments{
\item{data}{a numerical vector with vote shares}
}
\value{
a vector of rmps values
}
\description{
Computes the relative measure of party size as suggested by 
Giebler/Lacewell/Regel/Werner 2015.
}
\references{
Giebler, Heiko, Onawa Promise Lacewell, Sven Regel and Annika Werner. 2015. 
Niedergang oder Wandel? Parteitypen und die Krise der repraesentativen Demokratie. 
In Steckt die Demokratie in der Krise?, ed. Wolfgang Merkel, 181-219. Wiesbaden: Springer VS.
}
