% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaling_rile.R
\name{rile}
\alias{logit_rile}
\alias{rile}
\title{RILE}
\usage{
rile(x)

logit_rile(x)
}
\arguments{
\item{x}{A data.frame with cases to be scaled, variables named "per..."}

\item{...}{A ManifestoCorpus or ManifestoDocument with annotated texts to be be scaled}
}
\description{
Computes the RILE or other bipolar linear scaling measures for each case in a
data.frame or ManifestoCorpus
}

